Adapt abstract predicate validation to new spec checker structure
